sw_be_container

Module tokio_tools

Module to handle easy sending functions to tokio

The provides two functions one function run_in_tokio creates and sends the function to tokio. The second function run_in_tokio_with_cancel allows the creation of a CancellationToken which can be used to shut down the tokio async.

Structs

ThreadRuntime

Functions

rt_multithreaded

run_in_tokio

run async function inside tokio instance on current thread

run_in_tokio_with_cancel

Run async with cancellability via CancellationToken